requirements.txt
picamera2
opencv-python
numpy
pillow
keyboard